+++ app/code/Magento/Core/Model/Session/Config.php
@@ -174,7 +174,7 @@ class Config implements \Magento\Session\Config\ConfigInterface
         $domain = $this->_storeConfig->getConfig(self::XML_PATH_COOKIE_DOMAIN, $this->_storeManager->getStore());
         $domain = empty($domain) ? $this->_httpRequest->getHttpHost() : $domain;
         $this->setCookieDomain((string)$domain);
-        $this->setCookieSecure($this->_httpRequest->isSecure());
+
         $this->setCookieHttpOnly(
             $this->_storeConfig->getConfig(self::XML_PATH_COOKIE_HTTPONLY, $this->_storeManager->getStore())
         );